requirements.txt
appdirs==1.4.4
asgiref==3.9.1
attrs==25.3.0
autoslug==1.0.6
Django==5.2.4
django-autoslug==1.9.9
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
fs==2.4.16
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
setuptools==78.1.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
unicode-slugify==0.1.5
Unidecode==1.4.0
uritemplate==4.2.0
wheel==0.45.1
